Setup (#4)
* initial

* setup complete
