+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -8,37 +8,7 @@
 
 ## Reviewer Only
 
-- [ ] **Code Quality**
-  - [ ] Are variable and function names descriptive and follow naming conventions?
-  - [ ] Is the code free of redundant or duplicate segments?
-  - [ ] Does the code adhere to the project's style guidelines? Github linter checks shold pass.
-
-- [ ] **Functionality**
-  - [ ] Does the code perform the intended functions correctly?
-  - [ ] Are all possible edge cases handled appropriately?
-
-- [ ] **Testing**
-  - [ ] Are there unit tests covering new or modified functionalities?
-  - [ ] Do all existing and new tests pass successfully?
-  - [ ] Are Github checks passing?
-
-- [ ] **Documentation**
-  - [ ] Is the code adequately commented to explain complex sections?
-  - [ ] Have docstrings been used in functions/ methods?
-  - [ ] Do docstrings follow convention throughout solution?
-  - [ ] Has relevant documentation been updated or created?
-
-- [ ] **Dependencies**
-  - [ ] Are new dependencies justified and documented?
-  - [ ] Is the impact of these dependencies on the overall project considered?
-
-- [ ] **Performance**
-  - [ ] Is the code optimised for performance without compromising readability?
-
-- [ ] **Security**
-  - [ ] Are there any security implications introduced by the changes?
-  - [ ] Have security best practices been followed?
-
-- [ ] **Merge Preparations**
-  - [ ] Is the branch up-to-date with the base branch?
-  - [ ] Are there any merge conflicts that need resolution?
+- [] Github checks are passing
+- [] Unit tests are passing
+- [] Docstrings, comments, and other documentation added where necessary
+- [] Code performs the intended functionality
