--- .github/funding.yml
@@ -1,3 +1,0 @@
-# These are supported funding model platforms
-
-github: [Jordan-Prescott]
