+++ src/thors_hammer/__init__.py
@@ -1,3 +1,5 @@
 from .api import API as API
-from .scripter import Scripter as Scripter
-from .reporter import Reporter as Reporter
+from .agent import Agent as Agent
+from .bulks import Bulker as Bulker
+
+__all__ = ["API", "Agent", "Bulker"]
