+++ src/thors_hammer/exceptions.py
@@ -1,0 +1,1 @@
+# host all exceptions
