--- src/thors_hammer/reporter.py
@@ -1,6 +1,0 @@
-from .api import API
-
-
-class Reporter:
-    def __init__(self, api: API):
-        self.api = api
