+++ src/thors_hammer/requester.py
@@ -1,0 +1,4 @@
+# this will be resposnsible for sending and receiving data from the API
+
+class Requester:
+    pass
