+++ src/thors_hammer/utils/parser.py
@@ -1,0 +1,3 @@
+# Responsible for parsing data between types such as JSON, XML and Classes
+
+# Design not fully fleshed out yet
